unify InboundFrame and InprocPeerSnapshot into omq-proto
Both backends now share a single InboundFrame enum and
InprocPeerSnapshot struct defined in omq_proto::inproc. The
per-backend duplicates are removed. InboundFrame temporarily
retains the InboundMessage wrapper with peer_identity to keep
compio's recv path working; the next commit removes it.
